requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
httpx
python-dotenv
pytest
yandexcloud
cryptography
PyJWT